requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pdfplumber
pypdfium2
python-docx
docx2txt
crewai
openai
langchain-openai
